Refactor RDW crate into SRP modules
