+++ crates/copybook-rdw/src/lib.rs
@@ -10,10 +10,18 @@
 //! and [`RDWRecordWriter`] for streaming record I/O, or the lower-level helpers
 //! ([`rdw_read_len`], [`rdw_slice_body`]) for custom framing.
 
-use copybook_core::Schema;
-use copybook_error::{Error, ErrorCode, ErrorContext, Result};
-use std::io::{BufRead, BufReader, Read, Write};
-use tracing::{debug, warn};
+mod header;
+mod primitives;
+mod reader;
+mod record;
+mod schema_prefix;
+mod writer;
+
+pub use header::{RdwHeader, rdw_is_suspect_ascii_corruption, rdw_payload_len_to_u16};
+pub use primitives::{rdw_read_len, rdw_slice_body, rdw_try_peek_len, rdw_validate_and_finish};
+pub use reader::RDWRecordReader;
+pub use record::RDWRecord;
+pub use writer::RDWRecordWriter;
 
 /// Size of an RDW header in bytes.
 pub const RDW_HEADER_LEN: usize = 4;
@@ -23,777 +31,11 @@ pub const RDW_MAX_PAYLOAD_LEN: usize = u16::MAX as usize;
 
 const RDW_READER_BUF_CAPACITY: usize = (u16::MAX as usize) + RDW_HEADER_LEN;
 
-/// Parsed RDW header (`length + reserved`).
-///
-/// A 4-byte Record Descriptor Word containing a 2-byte big-endian payload
-/// length and 2 reserved bytes.
-///
-/// # Examples
-///
-/// ```
-/// use copybook_rdw::RdwHeader;
-///
-/// let header = RdwHeader::from_bytes([0x00, 0x50, 0x00, 0x00]);
-/// assert_eq!(header.length(), 80);
-/// assert_eq!(header.reserved(), 0);
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct RdwHeader {
-    bytes: [u8; RDW_HEADER_LEN],
-}
-
-impl RdwHeader {
-    /// Construct from raw 4-byte header bytes.
-    #[must_use]
-    #[inline]
-    pub const fn from_bytes(bytes: [u8; RDW_HEADER_LEN]) -> Self {
-        Self { bytes }
-    }
-
-    /// Construct from payload length and reserved bytes.
-    ///
-    /// # Errors
-    /// Returns `CBKF102_RECORD_LENGTH_INVALID` when `payload_len > u16::MAX`.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn from_payload_len(payload_len: usize, reserved: u16) -> Result<Self> {
-        let len = rdw_payload_len_to_u16(payload_len)?;
-        let len_bytes = len.to_be_bytes();
-        let reserved_bytes = reserved.to_be_bytes();
-        Ok(Self {
-            bytes: [
-                len_bytes[0],
-                len_bytes[1],
-                reserved_bytes[0],
-                reserved_bytes[1],
-            ],
-        })
-    }
-
-    /// Return raw bytes.
-    #[must_use]
-    #[inline]
-    pub const fn bytes(self) -> [u8; RDW_HEADER_LEN] {
-        self.bytes
-    }
-
-    /// Extract payload length.
-    #[must_use]
-    #[inline]
-    pub const fn length(self) -> u16 {
-        u16::from_be_bytes([self.bytes[0], self.bytes[1]])
-    }
-
-    /// Extract reserved bytes.
-    #[must_use]
-    #[inline]
-    pub const fn reserved(self) -> u16 {
-        u16::from_be_bytes([self.bytes[2], self.bytes[3]])
-    }
-
-    /// ASCII-corruption heuristic for the length bytes.
-    ///
-    /// Returns `true` when both length bytes look like ASCII digits.
-    #[must_use]
-    #[inline]
-    pub const fn looks_ascii_corrupt(self) -> bool {
-        rdw_is_suspect_ascii_corruption(self.bytes)
-    }
-}
-
-/// Convert payload length to `u16` with RDW error mapping.
-///
-/// # Errors
-/// Returns `CBKF102_RECORD_LENGTH_INVALID` when `len > u16::MAX`.
-#[inline]
-#[must_use = "Handle the Result or propagate the error"]
-pub fn rdw_payload_len_to_u16(len: usize) -> Result<u16> {
-    u16::try_from(len).map_err(|_| {
-        Error::new(
-            ErrorCode::CBKF102_RECORD_LENGTH_INVALID,
-            format!(
-                "RDW payload too large: {} bytes exceeds maximum of {}",
-                len,
-                u16::MAX
-            ),
-        )
-    })
-}
-
-/// Heuristic to detect ASCII-corrupted RDW headers.
-///
-/// Returns `true` when both RDW length bytes are ASCII digits (`0x30..=0x39`).
-#[must_use]
-#[inline]
-pub const fn rdw_is_suspect_ascii_corruption(rdw_header: [u8; RDW_HEADER_LEN]) -> bool {
-    copybook_rdw_predicates::rdw_is_suspect_ascii_corruption(rdw_header)
-}
-
-/// Read a 2-byte big-endian RDW body length and consume those two bytes.
-///
-/// # Errors
-/// Returns:
-/// - `CBKF104_RDW_SUSPECT_ASCII` for I/O errors while peeking.
-/// - `CBKF102_RECORD_LENGTH_INVALID` for incomplete headers.
-#[inline]
-#[must_use = "Handle the Result or propagate the error"]
-pub fn rdw_read_len<R: BufRead>(reader: &mut R) -> Result<u16> {
-    let buf = reader.fill_buf().map_err(|e| {
-        Error::new(
-            ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-            format!("I/O error peeking RDW length: {e}"),
-        )
-    })?;
-    if buf.len() < 2 {
-        return Err(Error::new(
-            ErrorCode::CBKF102_RECORD_LENGTH_INVALID,
-            format!(
-                "Incomplete RDW header: expected 2 bytes for length (have {})",
-                buf.len()
-            ),
-        ));
-    }
-
-    let hi = buf[0];
-    let lo = buf[1];
-    reader.consume(2);
-    Ok(u16::from_be_bytes([hi, lo]))
-}
-
-/// Borrow the RDW body slice for `len` bytes without consuming.
-///
-/// # Errors
-/// Returns:
-/// - `CBKF104_RDW_SUSPECT_ASCII` for I/O errors while peeking.
-/// - `CBKF102_RECORD_LENGTH_INVALID` when fewer than `len` bytes are available.
-#[inline]
-#[must_use = "Handle the Result or propagate the error"]
-pub fn rdw_slice_body<R: BufRead>(reader: &mut R, len: u16) -> Result<&[u8]> {
-    let need = usize::from(len);
-    if need == 0 {
-        return Ok(&[]);
-    }
-
-    let buf = reader.fill_buf().map_err(|e| {
-        Error::new(
-            ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-            format!("I/O error reading RDW payload: {e}"),
-        )
-    })?;
-
-    if buf.len() < need {
-        return Err(Error::new(
-            ErrorCode::CBKF102_RECORD_LENGTH_INVALID,
-            format!(
-                "Incomplete RDW record payload: expected {} bytes (have {})",
-                need,
-                buf.len()
-            ),
-        ));
-    }
-
-    Ok(&buf[..need])
-}
-
-/// Placeholder for future RDW body validation hooks.
-#[inline]
-#[must_use]
-pub const fn rdw_validate_and_finish(body: &[u8]) -> &[u8] {
-    body
-}
-
-/// Probe if enough bytes exist to attempt RDW length parsing.
-///
-/// - `0` or `1` byte buffered => `Ok(None)`
-/// - `>= 2` bytes buffered => `Ok(Some(()))`
-///
-/// # Errors
-/// Returns `CBKF104_RDW_SUSPECT_ASCII` for I/O errors while peeking.
-#[inline]
-#[must_use = "Handle the Result or propagate the error"]
-pub fn rdw_try_peek_len<R: BufRead>(reader: &mut R) -> Result<Option<()>> {
-    let buf = reader.fill_buf().map_err(|e| {
-        Error::new(
-            ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-            format!("I/O error peeking RDW header: {e}"),
-        )
-    })?;
-    if buf.len() <= 1 {
-        return Ok(None);
-    }
-    Ok(Some(()))
-}
-
-/// An RDW record with header and payload bytes.
-#[derive(Debug, Clone)]
-pub struct RDWRecord {
-    /// 4-byte RDW header (length + reserved).
-    pub header: [u8; RDW_HEADER_LEN],
-    /// Record payload bytes.
-    pub payload: Vec<u8>,
-}
-
-impl RDWRecord {
-    /// Create a new RDW record from payload (fallible constructor).
-    ///
-    /// Constructs an [`RDWRecord`] with a computed header (payload length + zero reserved bytes).
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use copybook_rdw::RDWRecord;
-    ///
-    /// let record = RDWRecord::try_new(vec![0xC8; 80]).unwrap();
-    /// assert_eq!(record.length(), 80);
-    /// assert_eq!(record.payload.len(), 80);
-    /// ```
-    ///
-    /// # Errors
-    /// Returns an error when payload length exceeds `u16::MAX`.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn try_new(payload: Vec<u8>) -> Result<Self> {
-        let header = RdwHeader::from_payload_len(payload.len(), 0)?.bytes();
-        Ok(Self { header, payload })
-    }
-
-    /// Create a new RDW record from payload.
-    ///
-    /// # Panics
-    /// Panics when payload length exceeds `u16::MAX`.
-    #[deprecated(
-        since = "0.4.3",
-        note = "use try_new() instead for fallible construction"
-    )]
-    #[allow(clippy::expect_used)] // Intentional panic for deprecated API
-    #[inline]
-    #[must_use]
-    pub fn new(payload: Vec<u8>) -> Self {
-        Self::try_new(payload).expect("RDW payload exceeds maximum size (65535 bytes)")
-    }
-
-    /// Create an RDW record preserving reserved bytes (fallible constructor).
-    ///
-    /// # Errors
-    /// Returns an error when payload length exceeds `u16::MAX`.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn try_with_reserved(payload: Vec<u8>, reserved: u16) -> Result<Self> {
-        let header = RdwHeader::from_payload_len(payload.len(), reserved)?.bytes();
-        Ok(Self { header, payload })
-    }
-
-    /// Create an RDW record preserving reserved bytes.
-    ///
-    /// # Panics
-    /// Panics when payload length exceeds `u16::MAX`.
-    #[deprecated(
-        since = "0.4.3",
-        note = "use try_with_reserved() instead for fallible construction"
-    )]
-    #[allow(clippy::expect_used)] // Intentional panic for deprecated API
-    #[inline]
-    #[must_use]
-    pub fn with_reserved(payload: Vec<u8>, reserved: u16) -> Self {
-        Self::try_with_reserved(payload, reserved)
-            .expect("RDW payload exceeds maximum size (65535 bytes)")
-    }
-
-    /// Get payload length from header.
-    #[inline]
-    #[must_use]
-    pub fn length(&self) -> u16 {
-        RdwHeader::from_bytes(self.header).length()
-    }
-
-    /// Get reserved bytes from header.
-    #[inline]
-    #[must_use]
-    pub fn reserved(&self) -> u16 {
-        RdwHeader::from_bytes(self.header).reserved()
-    }
-
-    /// Recompute the header length field from payload length.
-    ///
-    /// # Errors
-    /// Returns an error when payload length exceeds `u16::MAX`.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn try_recompute_length(&mut self) -> Result<()> {
-        self.header = RdwHeader::from_payload_len(self.payload.len(), self.reserved())?.bytes();
-        Ok(())
-    }
-
-    /// Recompute the header length field from payload length.
-    ///
-    /// # Panics
-    /// Panics when payload length exceeds `u16::MAX`.
-    #[deprecated(
-        since = "0.4.3",
-        note = "use try_recompute_length() instead for fallible operation"
-    )]
-    #[allow(clippy::expect_used)] // Intentional panic for deprecated API
-    #[inline]
-    pub fn recompute_length(&mut self) {
-        self.try_recompute_length()
-            .expect("RDW payload exceeds maximum size (65535 bytes)");
-    }
-
-    /// Serialize record as `header + payload`.
-    #[inline]
-    #[must_use]
-    pub fn as_bytes(&self) -> Vec<u8> {
-        let mut result = Vec::with_capacity(RDW_HEADER_LEN + self.payload.len());
-        result.extend_from_slice(&self.header);
-        result.extend_from_slice(&self.payload);
-        result
-    }
-}
-
-/// RDW (Record Descriptor Word) record reader for variable-length records.
-#[derive(Debug)]
-pub struct RDWRecordReader<R: Read> {
-    input: BufReader<R>,
-    record_count: u64,
-    strict_mode: bool,
-}
-
-impl<R: Read> RDWRecordReader<R> {
-    /// Create a new RDW record reader.
-    #[inline]
-    #[must_use]
-    pub fn new(input: R, strict_mode: bool) -> Self {
-        Self {
-            input: BufReader::with_capacity(RDW_READER_BUF_CAPACITY, input),
-            record_count: 0,
-            strict_mode,
-        }
-    }
-
-    #[inline]
-    fn peek_header(&mut self) -> Result<Option<[u8; RDW_HEADER_LEN]>> {
-        let peek = rdw_try_peek_len(&mut self.input).map_err(|error| {
-            error.with_context(ErrorContext {
-                record_index: Some(self.record_count + 1),
-                field_path: None,
-                byte_offset: Some(0),
-                line_number: None,
-                details: Some("Unable to peek RDW header".to_string()),
-            })
-        })?;
-
-        if peek.is_none() {
-            let buf = self.input.fill_buf().map_err(|e| {
-                Error::new(
-                    ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-                    format!("I/O error reading RDW header: {e}"),
-                )
-                .with_context(ErrorContext {
-                    record_index: Some(self.record_count + 1),
-                    field_path: None,
-                    byte_offset: Some(0),
-                    line_number: None,
-                    details: Some("Unable to read RDW header".to_string()),
-                })
-            })?;
-
-            if buf.is_empty() {
-                debug!("Reached EOF after {} RDW records", self.record_count);
-                return Ok(None);
-            }
-
-            if self.strict_mode {
-                return Err(Error::new(
-                    ErrorCode::CBKF221_RDW_UNDERFLOW,
-                    "Incomplete RDW header: expected 4 bytes".to_string(),
-                )
-                .with_context(ErrorContext {
-                    record_index: Some(self.record_count + 1),
-                    field_path: None,
-                    byte_offset: Some(0),
-                    line_number: None,
-                    details: Some("File ends with incomplete RDW header".to_string()),
-                }));
-            }
-
-            debug!(
-                "Reached EOF after {} RDW records (truncated header ignored)",
-                self.record_count
-            );
-            let remaining = buf.len();
-            self.input.consume(remaining);
-            return Ok(None);
-        }
-
-        let buf = self.input.fill_buf().map_err(|e| {
-            Error::new(
-                ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-                format!("I/O error reading RDW header: {e}"),
-            )
-            .with_context(ErrorContext {
-                record_index: Some(self.record_count + 1),
-                field_path: None,
-                byte_offset: Some(0),
-                line_number: None,
-                details: Some("Unable to read RDW header".to_string()),
-            })
-        })?;
-
-        if buf.len() < RDW_HEADER_LEN {
-            if self.strict_mode {
-                return Err(Error::new(
-                    ErrorCode::CBKF221_RDW_UNDERFLOW,
-                    "Incomplete RDW header: expected 4 bytes".to_string(),
-                )
-                .with_context(ErrorContext {
-                    record_index: Some(self.record_count + 1),
-                    field_path: None,
-                    byte_offset: Some(0),
-                    line_number: None,
-                    details: Some("File ends with incomplete RDW header".to_string()),
-                }));
-            }
-
-            debug!(
-                "Reached EOF after {} RDW records (truncated header ignored)",
-                self.record_count
-            );
-            let remaining = buf.len();
-            self.input.consume(remaining);
-            return Ok(None);
-        }
-
-        Ok(Some([buf[0], buf[1], buf[2], buf[3]]))
-    }
-
-    /// Read the next RDW record.
-    ///
-    /// # Errors
-    /// Returns an error if the record cannot be read due to I/O errors or
-    /// framing issues.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn read_record(&mut self) -> Result<Option<RDWRecord>> {
-        let Some(header) = self.peek_header()? else {
-            return Ok(None);
-        };
-
-        let length = match rdw_read_len(&mut self.input) {
-            Ok(len) => len,
-            Err(error) => {
-                return Err(error.with_context(ErrorContext {
-                    record_index: Some(self.record_count + 1),
-                    field_path: None,
-                    byte_offset: Some(0),
-                    line_number: None,
-                    details: Some("Unable to read RDW body length".to_string()),
-                }));
-            }
-        };
-
-        // Consume reserved bytes so the buffer now points at the body.
-        self.input.consume(2);
-        let reserved = u16::from_be_bytes([header[2], header[3]]);
-
-        self.record_count += 1;
-        debug!(
-            "Read RDW header for record {}: length={}, reserved={:04X}",
-            self.record_count,
-            u32::from(length),
-            reserved
-        );
-
-        if reserved != 0 {
-            let error = Error::new(
-                ErrorCode::CBKR211_RDW_RESERVED_NONZERO,
-                format!("RDW reserved bytes are non-zero: {reserved:04X}"),
-            )
-            .with_context(ErrorContext {
-                record_index: Some(self.record_count),
-                field_path: None,
-                byte_offset: Some(2),
-                line_number: None,
-                details: Some(format!("Expected 0000, got {reserved:04X}")),
-            });
-
-            if self.strict_mode {
-                return Err(error);
-            }
-
-            warn!(
-                "RDW reserved bytes non-zero (record {}): {:04X}",
-                self.record_count, reserved
-            );
-        }
-
-        if Self::is_suspect_ascii_corruption(header) {
-            warn!(
-                "RDW appears to be ASCII-corrupted (record {}): {:02X} {:02X} {:02X} {:02X}",
-                self.record_count, header[0], header[1], header[2], header[3]
-            );
-
-            return Err(Error::new(
-                ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-                format!(
-                    "RDW appears to be ASCII-corrupted: {:02X} {:02X} {:02X} {:02X}",
-                    header[0], header[1], header[2], header[3]
-                ),
-            )
-            .with_context(ErrorContext {
-                record_index: Some(self.record_count),
-                field_path: None,
-                byte_offset: Some(0),
-                line_number: None,
-                details: Some("Suspected ASCII transfer corruption".to_string()),
-            }));
-        }
-
-        if length == 0 {
-            debug!("Zero-length RDW record {}", self.record_count);
-            return Ok(Some(RDWRecord {
-                header,
-                payload: Vec::new(),
-            }));
-        }
-
-        let payload_len = usize::from(length);
-        let body_slice = match rdw_slice_body(&mut self.input, length) {
-            Ok(slice) => slice,
-            Err(error) => {
-                return Err(error.with_context(ErrorContext {
-                    record_index: Some(self.record_count),
-                    field_path: None,
-                    byte_offset: Some(4),
-                    line_number: None,
-                    details: Some("File ends with incomplete RDW payload".to_string()),
-                }));
-            }
-        };
-
-        let payload = rdw_validate_and_finish(body_slice).to_vec();
-        self.input.consume(payload_len);
-
-        debug!(
-            "Read RDW record {} payload: {} bytes",
-            self.record_count, length
-        );
-        Ok(Some(RDWRecord { header, payload }))
-    }
-
-    /// Validate a zero-length record against schema requirements.
-    ///
-    /// # Errors
-    /// Returns an error when the schema requires non-zero fixed bytes.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn validate_zero_length_record(&self, schema: &Schema) -> Result<()> {
-        let min_size = Self::calculate_schema_fixed_prefix(schema);
-
-        if min_size > 0 {
-            return Err(Error::new(
-                ErrorCode::CBKF221_RDW_UNDERFLOW,
-                format!("Zero-length RDW record invalid: schema requires minimum {min_size} bytes"),
-            )
-            .with_context(ErrorContext {
-                record_index: Some(self.record_count),
-                field_path: None,
-                byte_offset: None,
-                line_number: None,
-                details: Some("Zero-length record with non-zero schema prefix".to_string()),
-            }));
-        }
-
-        Ok(())
-    }
-
-    /// Number of RDW records consumed from the stream.
-    #[inline]
-    #[must_use]
-    pub fn record_count(&self) -> u64 {
-        self.record_count
-    }
-
-    #[inline]
-    fn calculate_schema_fixed_prefix(schema: &Schema) -> u32 {
-        use copybook_core::{Field, Occurs};
-
-        fn find_first_odo_offset(fields: &[Field], current: &mut Option<u32>) {
-            for field in fields {
-                if let Some(Occurs::ODO { .. }) = &field.occurs {
-                    let offset = field.offset;
-                    match current {
-                        Some(existing) => {
-                            if offset < *existing {
-                                *current = Some(offset);
-                            }
-                        }
-                        None => *current = Some(offset),
-                    }
-                }
-                if !field.children.is_empty() {
-                    find_first_odo_offset(&field.children, current);
-                }
-            }
-        }
-
-        let mut first_odo_offset: Option<u32> = None;
-        find_first_odo_offset(&schema.fields, &mut first_odo_offset);
-
-        if let Some(offset) = first_odo_offset {
-            offset
-        } else if let Some(lrecl) = schema.lrecl_fixed {
-            lrecl
-        } else {
-            fn find_record_end(fields: &[Field], max_end: &mut u32) {
-                for field in fields {
-                    let end = field.offset + field.len;
-                    if end > *max_end {
-                        *max_end = end;
-                    }
-                    if !field.children.is_empty() {
-                        find_record_end(&field.children, max_end);
-                    }
-                }
-            }
-
-            let mut max_end = 0;
-            find_record_end(&schema.fields, &mut max_end);
-            max_end
-        }
-    }
-
-    #[inline]
-    fn is_suspect_ascii_corruption(rdw_header: [u8; RDW_HEADER_LEN]) -> bool {
-        rdw_is_suspect_ascii_corruption(rdw_header)
-    }
-}
-
-/// RDW record writer for variable-length records.
-#[derive(Debug)]
-pub struct RDWRecordWriter<W: Write> {
-    output: W,
-    record_count: u64,
-}
-
-impl<W: Write> RDWRecordWriter<W> {
-    /// Create a new RDW record writer.
-    #[inline]
-    #[must_use]
-    pub fn new(output: W) -> Self {
-        Self {
-            output,
-            record_count: 0,
-        }
-    }
-
-    /// Write an RDW record.
-    ///
-    /// # Errors
-    /// Returns an error if writing header or payload fails.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn write_record(&mut self, record: &RDWRecord) -> Result<()> {
-        self.output.write_all(&record.header).map_err(|e| {
-            Error::new(
-                ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-                format!("I/O error writing RDW header: {e}"),
-            )
-            .with_context(ErrorContext {
-                record_index: Some(self.record_count + 1),
-                field_path: None,
-                byte_offset: None,
-                line_number: None,
-                details: None,
-            })
-        })?;
-
-        self.output.write_all(&record.payload).map_err(|e| {
-            Error::new(
-                ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-                format!("I/O error writing RDW payload: {e}"),
-            )
-            .with_context(ErrorContext {
-                record_index: Some(self.record_count + 1),
-                field_path: None,
-                byte_offset: Some(4),
-                line_number: None,
-                details: None,
-            })
-        })?;
-
-        self.record_count += 1;
-        debug!(
-            "Wrote RDW record {} with {} byte payload",
-            self.record_count,
-            record.payload.len()
-        );
-        Ok(())
-    }
-
-    /// Write an RDW record directly from payload.
-    ///
-    /// # Errors
-    /// Returns an error if payload length exceeds `u16::MAX` or I/O fails.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn write_record_from_payload(
-        &mut self,
-        payload: &[u8],
-        preserve_reserved: Option<u16>,
-    ) -> Result<()> {
-        let length = payload.len();
-        let header =
-            RdwHeader::from_payload_len(length, preserve_reserved.unwrap_or(0)).map_err(|_| {
-                Error::new(
-                    ErrorCode::CBKE501_JSON_TYPE_MISMATCH,
-                    format!(
-                        "RDW payload too large: {length} bytes exceeds maximum of {}",
-                        u16::MAX
-                    ),
-                )
-                .with_context(ErrorContext {
-                    record_index: Some(self.record_count + 1),
-                    field_path: None,
-                    byte_offset: None,
-                    line_number: None,
-                    details: Some("RDW length field is 16-bit".to_string()),
-                })
-            })?;
-
-        let record = RDWRecord {
-            header: header.bytes(),
-            payload: payload.to_vec(),
-        };
-        self.write_record(&record)
-    }
-
-    /// Flush writer output.
-    ///
-    /// # Errors
-    /// Returns an error when flush fails.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn flush(&mut self) -> Result<()> {
-        self.output.flush().map_err(|e| {
-            Error::new(
-                ErrorCode::CBKF104_RDW_SUSPECT_ASCII,
-                format!("I/O error flushing output: {e}"),
-            )
-        })
-    }
-
-    /// Number of written RDW records.
-    #[inline]
-    #[must_use]
-    pub fn record_count(&self) -> u64 {
-        self.record_count
-    }
-}
-
 #[cfg(test)]
 #[allow(clippy::expect_used, clippy::unwrap_used, clippy::panic)]
 mod tests {
     use super::*;
+    use copybook_error::ErrorCode;
     use proptest::collection::vec;
     use proptest::prelude::*;
     use std::io::{BufRead, Cursor};
